2025-04-02 회사양식수정삭제

+++ group/src/main/java/com/yedam/app/group/service/ApprovalFormVO.java
@@ -10,7 +10,7 @@ public class ApprovalFormVO {
 	private String formType;    // 양식유형
 	private String remarks;     // 비고
 	private String content;     // 내용
-	private int version;        // 버전
+	private double version;        // 버전
 	private String active;      // 활성화 여부
 	private Date creationDate;  // 생성일
 	private int suberNo;        // 회사번호
